charts hotfix
